Add note about tag navigation
